requirements.txt
certifi==2023.5.7
mysql-connector==2.2.9
numpy==1.25.1
pandas==2.0.3
pycurl==7.45.2
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
